backend/pyproject.toml
[tool.poetry]
name = "djangoproduct"
version = "0.1.0"
description = ""
authors = ["Сергей Коновалов <SokiZNosoK68rus@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
isort = "^5.13.2"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"
ipython = "^8.31.0"
coverage = "^7.6.10"
djoser = "^2.3.1"
drf-spectacular = "^0.28.0"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
